[CORE-3931] use back previously created one
